refactor: move CopyrightNotice
