+++ src/components/TeacherComponents/CopyrightNotice/CopyrightNotice.test.tsx
@@ -1,8 +1,9 @@
 import { screen } from "@testing-library/react";
 
-import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
-
 import CopyrightNotice from "./CopyrightNotice";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 describe("CopyrightNotice", () => {
   it("renders pre-ALB copyright notice", () => {
