+++ src/components/TeacherComponents/CopyrightNotice/CopyrightNotice.tsx
