+++ src/components/TeacherComponents/CopyrightNotice/index.ts
